# web/src/shared/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Tamandua",
    "version": "0.9"
  },
  "paths": {
...
    "/api/policies/cra": {
      "get": {
        "tags": [
          "compliance"
        ],
        "summary": "Cra Policy",
        "description": "Anyone can read it (the panel shows or hides the CRA kit with it); an administrator changes it.",
        "operationId": "cra_policy_api_policies_cra_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CraPolicy"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "compliance"
        ],
        "summary": "Set Cra Policy",
        "description": "Turns the CRA kit on or off, with a reason kept in the history. Off keeps products and decisions.",
        "operationId": "set_cra_policy_api_policies_cra_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CraPolicyIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CraPolicy"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/cra": {
      "get": {
        "tags": [
          "compliance"
        ],
        "summary": "Overview",
        "operationId": "overview_api_cra_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CraOverview"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "compliance"
        ],
        "summary": "Change",
        "description": "Marks products, records the assessment of an event (only \"exploited in our product\" starts the clocks) and\nwhich stages were sent. Answers the new overview.",
        "operationId": "change_api_cra_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "oneOf": [
                  {
                    "$ref": "#/components/schemas/ProductIn"
                  },
                  {
                    "$ref": "#/components/schemas/UnproductIn"
                  },
                  {
                    "$ref": "#/components/schemas/MarkIn"
                  },
                  {
                    "$ref": "#/components/schemas/AssessIn"
                  },
                  {
                    "$ref": "#/components/schemas/ReopenIn"
                  }
                ],
                "title": "Body",
                "discriminator": {
                  "propertyName": "op",
                  "mapping": {
                    "product": "#/components/schemas/ProductIn",
                    "unproduct": "#/components/schemas/UnproductIn",
                    "mark": "#/components/schemas/MarkIn",
                    "assess": "#/components/schemas/AssessIn",
                    "reopen": "#/components/schemas/ReopenIn"
                  }
                }
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CraOverview"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/cra/products": {
      "get": {
        "tags": [
          "compliance"
        ],
        "summary": "Products",
        "operationId": "products_api_cra_products_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10000,
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CraProductPage"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/cra/events": {
      "get": {
        "tags": [
          "compliance"
        ],
        "summary": "Events",
        "description": "Most urgent first. The ENISA draft is only built for the exploited events on the page.",
        "operationId": "events_api_cra_events_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 10,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10000,
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CraEventPage"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/cra/assets": {
      "get": {
        "tags": [
          "compliance"
        ],
        "summary": "Assets",
        "description": "Analyzed assets that can still be marked as products, filtered by name.",
        "operationId": "assets_api_cra_assets_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 100,
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 25,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10000,
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CraAssetPage"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/evidence/assets": {
      "get": {
        "tags": [
          "compliance"
        ],
        "summary": "Evidence Assets",
        "description": "The asset picker: analyzed assets by name, with whether each has what an SBOM needs.",
        "operationId": "evidence_assets_api_evidence_assets_get",
        "parameters": [
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 100,
              "default": "",
              "title": "Q"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 10000,
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EvidenceAssetPage"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/evidence/portfolio": {
      "post": {
        "tags": [
          "compliance"
        ],
        "summary": "Portfolio Evidence",
        "description": "Consolidated audit evidence of every analyzed asset (open, fixed and exceptions), for one framework.",
        "operationId": "portfolio_evidence_api_evidence_portfolio_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PortfolioEvidenceIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Audit evidence (PDF) of every analyzed asset",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "maxLength": 50000000,
                  "format": "binary"
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/evidence/portfolio/sbom": {
      "get": {
        "tags": [
          "compliance"
        ],
        "summary": "Portfolio Sbom",
        "description": "One CycloneDX document: each asset a top-level component with its packages. `organization` names the portfolio.",
        "operationId": "portfolio_sbom_api_evidence_portfolio_sbom_get",
        "parameters": [
          {
            "name": "organization",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "maxLength": 120,
              "default": "",
              "title": "Organization"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "CycloneDX SBOM of every asset with a completed full scan",
            "content": {
              "application/vnd.cyclonedx+json": {
                "schema": {
                  "type": "string",
                  "format": "binary",
                  "maxLength": 100000000
                }
              }
            }
          },
          "400": {
            "description": "Invalid parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "AssessIn": {
        "properties": {
          "op": {
            "type": "string",
            "const": "assess",
            "title": "Op"
          },
          "event": {
            "type": "string",
            "maxLength": 300,
            "title": "Event"
          },
          "verdict": {
            "type": "string",
            "enum": [
              "not_affected",
              "exploited"
            ],
            "title": "Verdict"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 500
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "op",
          "event",
          "verdict"
        ],
        "title": "AssessIn"
      },
...
      "CraAsset": {
        "properties": {
          "key": {
            "type": "string",
            "title": "Key"
          },
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "key",
          "name"
        ],
        "title": "CraAsset"
      },
      "CraAssetPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CraAsset"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "CraAssetPage"
      },
...
      "CraEventPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CraEvent"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "CraEventPage"
      },
...
      "CraOverview": {
        "properties": {
          "reporting_page": {
            "type": "string",
            "title": "Reporting Page"
          },
          "counts": {
            "$ref": "#/components/schemas/CraCounts"
          }
        },
        "type": "object",
        "required": [
          "reporting_page",
          "counts"
        ],
        "title": "CraOverview"
      },
      "CraPolicy": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "by": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "By"
          },
          "at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "At"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          },
          "history": {
            "items": {
              "$ref": "#/components/schemas/CraPolicyChange"
            },
            "type": "array",
            "maxItems": 20,
            "title": "History"
          }
        },
        "type": "object",
        "required": [
          "enabled",
          "by",
          "at",
          "reason",
          "history"
        ],
        "title": "CraPolicy"
      },
...
      "CraPolicyIn": {
        "properties": {
          "enabled": {
            "type": "boolean",
            "title": "Enabled"
          },
          "reason": {
            "type": "string",
            "maxLength": 500,
            "title": "Reason"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "enabled",
          "reason"
        ],
        "title": "CraPolicyIn"
      },
...
      "CraProductPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/CraProduct"
            },
            "type": "array",
            "maxItems": 100,
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "limit": {
            "type": "integer",
            "title": "Limit"
          },
          "offset": {
            "type": "integer",
            "title": "Offset"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "CraProductPage"
      },
...
      "MarkIn": {
        "properties": {
          "op": {
            "type": "string",
            "const": "mark",
            "title": "Op"
          },
          "event": {
            "type": "string",
            "maxLength": 300,
            "title": "Event"
          },
          "stage": {
            "type": "string",
            "enum": [
              "early_warning",
              "notification",
              "final_report"
            ],
            "title": "Stage"
          },
          "sent": {
            "type": "boolean",
            "title": "Sent"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "op",
          "event",
          "stage",
          "sent"
        ],
        "title": "MarkIn"
      },
...
      "ProductIn": {
        "properties": {
          "op": {
            "type": "string",
            "const": "product",
            "title": "Op"
          },
          "key": {
            "type": "string",
            "maxLength": 200,
            "title": "Key"
          },
          "name": {
            "type": "string",
            "maxLength": 200,
            "title": "Name"
          },
          "support_until": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 10
              },
              {
                "type": "null"
              }
            ],
            "title": "Support Until"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "op",
          "key",
          "name",
          "support_until"
        ],
        "title": "ProductIn"
      },
...
      "ReopenIn": {
        "properties": {
          "op": {
            "type": "string",
            "const": "reopen",
            "title": "Op"
          },
          "event": {
            "type": "string",
            "maxLength": 300,
            "title": "Event"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "op",
          "event"
        ],
        "title": "ReopenIn"
      },
...
      "UnproductIn": {
        "properties": {
          "op": {
            "type": "string",
            "const": "unproduct",
            "title": "Op"
          },
          "key": {
            "type": "string",
            "maxLength": 200,
            "title": "Key"
          }
        },
        "additionalProperties": false,
        "type": "object",
        "required": [
          "op",
          "key"
        ],
        "title": "UnproductIn"
      },
...
      "Error": {
        "title": "Error",
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "type": "string",
            "title": "Error",
            "description": "In the reader's language."
          }
        }
      }
    },
    "securitySchemes": {
      "session": {
        "type": "apiKey",
        "in": "cookie",
        "name": "tamandua_session",
        "description": "Session signed in to the panel."
      },
      "metrics": {
        "type": "http",
        "scheme": "bearer",
        "description": "TAMANDUA_METRICS_TOKEN (Prometheus scraper)."
      }
    }
  },
  "security": [
    {
      "session": []
    }
  ]
}